removed operations and view.stream -> view.publisher
